moved constants file into components folder and moved components folder out of the pages folder
