+++ components/PlayerInfo.js
